Move "isWorkspaceFullScreen" back to the workspace state in the selector
