+++ src/state/selectors/index.js
@@ -3,3 +3,4 @@ export * from './canvases';
 export * from './config';
 export * from './manifests';
 export * from './windows';
+export * from './workspace';
