+++ src/state/selectors/workspace.js
@@ -1,0 +1,11 @@
+import { createSelector } from 'reselect';
+
+/** */
+function getWorkspace(state) {
+  return state.workspace;
+}
+
+export const getFullScreenEnabled = createSelector(
+  [getWorkspace],
+  workspace => workspace.isFullscreenEnabled,
+);
